+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.5.6] - 2024-01-08
+
 ### Added
 
 - bigquery: Users can now specify an alternate job project, for example for billing purposes. See docs for details.